# src/routes/documentation/openapi.json
{
	"openapi": "3.0.2",
	"info": {
		"title": "Shopping Cart API",
		"description": "Describes the API for Shopping Cart application.",
		"contact": {
			"name": "Sidnei Lechado",
			"email": "sidneilechado.dev@gmail.com"
		},
		"version": "1.0.0"
	},
...
	"paths": {
		"/api/user/create": {
			"post": {
				"operationId": "createUser",
				"summary": "Creates an user",
				"tags": [
					"user"
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"cpf": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "User successfully created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#components/schemas/User"
								}
							}
						}
					}
				}
			}
		},
...
		"/api/store/create": {
			"post": {
				"operationId": "createStore",
				"summary": "Creates a store",
				"tags": [
					"store"
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Store successfully created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#components/schemas/Store"
								}
							}
						}
					}
				}
			}
		},
		"/api/store/{id}": {
			"post": {
				"operationId": "getStore",
				"summary": "Get information from a particular store",
				"tags": [
					"store"
				],
				"parameters": [
					{
						"name": "id",
						"in": "params",
						"description": "Store id to be retrieved",
						"required": true,
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Store successfully retrieved",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#components/schemas/Store"
								}
							}
						}
					},
					"400": {
						"$ref": "#components/responses/InvalidData"
					}
				}
			}
		},
		"/api/product/create": {
			"post": {
				"operationId": "createProduct",
				"summary": "Creates a product",
				"tags": [
					"product"
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"price": {
										"type": "number"
									},
									"stockQuantity": {
										"type": "number"
									},
									"storeId": {
										"type": "number"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Product successfully created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#components/schemas/Product"
								}
							}
						}
					}
				}
			}
		},
...
	},
...
}